+++ packages/common/src/types/errors.ts
@@ -764,6 +764,20 @@ export class BigqueryTokenError extends LightdashError {
     }
 }
 
+/* This specific error will be used in the frontend
+to show a "reauthenticate" button in the UI
+*/
+export class RedshiftIamTokenError extends LightdashError {
+    constructor(message: string) {
+        super({
+            message,
+            name: 'RedshiftIamTokenError',
+            statusCode: 401,
+            data: {},
+        });
+    }
+}
+
 export class CustomSqlQueryForbiddenError extends LightdashError {
     constructor(
         message: string = 'User cannot run queries with custom SQL dimensions',